+++ CHANGELOG.md
@@ -12,6 +12,11 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 * Add `__json__` to `kwimage.Boxes`
 * Allow user to specify backend order in `load_image_shape`.
 
+### Changed
+* Demodata is now more robust to networking outages. If no mirror can be
+  reached, it generates random data to replace the real image.
+
+
 
 ## Version 0.9.25 - Released 2024-03-19